Breaking out the generate command into a separate file
I think this makes it easier to reason about as subcommands.
